thermal_prediction_project/requirements.txt
# Predictive Thermal Management System - Requirements
# Python 3.8 or higher required

# Core Data Science Libraries
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# System Monitoring
psutil>=5.8.0

# Model Persistence
joblib>=1.1.0

# Serial Communication (for Arduino)
pyserial>=3.5

# Optional: Jupyter for analysis
jupyter>=1.0.0
ipython>=7.30.0

# Optional: Advanced ML libraries
xgboost>=1.5.0  # Gradient boosting
lightgbm>=3.3.0  # Fast gradient boosting
tensorflow>=2.8.0  # Deep learning (optional)

# Optional: Web interface
flask>=2.0.0  # If building web dashboard
plotly>=5.5.0  # Interactive plots

# Development tools
pytest>=7.0.0  # Testing
black>=22.0.0  # Code formatting
flake8>=4.0.0  # Linting

# Installation Instructions:
# 
# 1. Basic installation (required):
#    pip install pandas numpy scikit-learn matplotlib seaborn psutil joblib pyserial
#
# 2. Full installation (with optional packages):
#    pip install -r requirements.txt
#
# 3. For Conda users:
#    conda install pandas numpy scikit-learn matplotlib seaborn psutil joblib
#    pip install pyserial
#
# 4. Verify installation:
#    python -c "import pandas, numpy, sklearn, matplotlib, psutil; print('Success!')"
